到第七章以前

+++ 0台大系統資訊訓練班/02進階班/004 第四章數字加法計算機完成檔案/第四章數字加法計算機完成檔案/main.js
@@ -1,4 +1,5 @@
 // 取得畫面上要控制的標籤(DOM, Document Object Model)
+// 好幾個常數後面用逗號
 const form = document.getElementById("form"),
     xInput = document.getElementById("xInput"),
     yInput = document.getElementById("yInput"),
@@ -10,7 +11,10 @@ form.addEventListener("submit", function (e) {
     e.preventDefault();
     // 表單送出後的流程
     console.log("表單被送出囉!", e);
-    const x = parseInt(xInput.value);
+
+
+    // const x = parseInt(xInput.value);
+    const x = parseFloat(xInput.value);
     // console.log("x", x);
     const y = parseInt(yInput.value);
     // console.log("y", y);
@@ -32,10 +36,14 @@ form.addEventListener("submit", function (e) {
 // const a = 30;
 // const b = 90;
 // const c = "90";
+
+// 轉換為整數或小數
 // parseInt("30.5") => 30
 // paserFloat("30.5") => 30.5
 // console.log(a + b);
 // console.log(a + parseInt(c));
+
+
 // Ctrl + /
 // console.log(a - b);
 // console.log(a * b);
